requirements.txt
pypdf==4.0.2
reportlab==4.1.0
PyPDF2==3.0.1
img2pdf==0.5.1
scikit-learn==1.4.1.post1
numpy==1.26.4
matplotlib==3.8.3
pillow==10.2.0
pdfminer==20191125
